refactor socket and create test users

+++ src/user/user.service.ts
@@ -18,20 +18,37 @@ export class UserService {
     return this.userRepository.findOneBy({ id: userId });
   }
 
+  async getUserByEmail(email: string) {
+    return this.userRepository.findOneBy({ email });
+  }
+
   async createUser(user: CreateUserRequest) {
     return this.userRepository.save({ ...user });
   }
 
   async deleteUser(userId: string) {
     const user = await this.getUser(userId);
-    const userProfile = await user.profile;
-    await this.userRepository.update(userId, {
-      profile: null,
-    });
+    if (!user) {
+      return false;
+    }
 
-    const removedProfile = await this.userProfileRepository.remove(userProfile);
-    const removedUser = await this.userRepository.remove(user);
-    return !!removedUser && !!removedProfile;
+    const userProfile = await user.profile;
+    if (userProfile) {
+      await this.userRepository.update(userId, {
+        profile: null,
+      });
+      await this.userProfileRepository.remove(userProfile);
+    }
+    await this.userRepository.remove(user);
+    return true;
+  }
+
+  async deleteUserByEmail(email: string) {
+    const user = await this.getUserByEmail(email);
+    if (!user) {
+      return false;
+    }
+    return this.deleteUser(user.id);
   }
 
   async isUserExistByEmail(email: string) {
